+++ client/src/components/Appointment/Appointment.styles.js
@@ -1,0 +1,11 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+
+
+export const useStyles = makeStyles({
+    root: {    
+        flexDirection: 'row',
+        justifyContent: 'center',
+        height: "100vh"
+    }
+});
