+++ client/src/pages/Appointments/Appointments.jsx
@@ -1,16 +1,15 @@
 import 'date-fns';
 import React from 'react';
 import Grid from '@material-ui/core/Grid';
-import { AppointmentHeader } from '../../components';
+import  Header  from '../../components/Appointment/Header';
 
 export default function Appointment() {
     return (
             <Grid container item xs={12}
-                container
                 direction="column"
                 justify="space-between"
                 alignItems="center">
-                <AppointmentHeader />
+                <Header />
             </Grid>
     );
 }
